+++ cmd/note.go
@@ -39,6 +39,9 @@ var noteViewCmd = &cobra.Command{
 }
 
 func init() {
+
+	noteCmd.Aliases = append(contactCmd.Aliases, "notes")
+
 	RootCmd.AddCommand(noteCmd)
 
 	noteCmd.AddCommand(noteAddCmd)
@@ -52,6 +55,8 @@ func init() {
 	noteAddCmd.Flags().Int("id", 0, "id of contact")
 	noteAddCmd.MarkFlagRequired("id")
 
+	noteViewCmd.Flags().BoolP("raw", "r", false, "show the raw output")
+
 }
 
 func noteFlags(cmd *cobra.Command) {
@@ -60,5 +65,6 @@ func noteFlags(cmd *cobra.Command) {
 	cmd.Flags().BoolP("call", "c", false, "this is/was a call")
 	cmd.Flags().BoolP("email", "e", false, "this is/was a email")
 	cmd.Flags().String("header", "", "header for this note")
+	cmd.Flags().BoolP("fast", "f", false, "create a quick note")
 
 }